+++ packages/cosmwasm/src/index.ts
@@ -21,7 +21,7 @@ export {
 } from "./cosmwasmclient";
 export {
   ExecuteResult,
-  FeeTable,
+  CosmWasmFeeTable,
   InstantiateOptions,
   InstantiateResult,
   MigrateResult,