+++ packages/faucet/src/constants.ts
@@ -1,12 +1,14 @@
-import { Coin, parseCoins } from "@cosmjs/launchpad";
+import { CosmosFeeTable, GasLimits, GasPrice } from "@cosmjs/launchpad";
 
 import { TokenConfiguration } from "./tokenmanager";
 import { parseBankTokens } from "./tokens";
 
 export const binaryName = "cosmos-faucet";
 export const memo: string | undefined = process.env.FAUCET_MEMO;
-export const fee: readonly Coin[] = parseCoins(process.env.FAUCET_FEE || "2000ucosm");
-export const gas: string = process.env.FAUCET_GAS || "80000";
+export const gasPrice = GasPrice.fromString(process.env.FAUCET_GAS_PRICE || "0.025ucosm");
+export const gasLimits: GasLimits<CosmosFeeTable> = {
+  send: parseInt(process.env.FAUCET_GAS_LIMIT || "80000", 10),
+};
 export const concurrency: number = Number.parseInt(process.env.FAUCET_CONCURRENCY || "", 10) || 5;
 export const port: number = Number.parseInt(process.env.FAUCET_PORT || "", 10) || 8000;
 export const mnemonic: string | undefined = process.env.FAUCET_MNEMONIC;