+++ packages/launchpad/src/signingcosmosclient.ts
@@ -1,31 +1,33 @@
 /* eslint-disable @typescript-eslint/naming-convention */
-import { Coin, coins } from "./coins";
+import { Coin } from "./coins";
 import { Account, BroadcastTxResult, CosmosClient, GetSequenceResult } from "./cosmosclient";
 import { makeSignBytes } from "./encoding";
+import { buildFeeTable, FeeTable, GasLimits, GasPrice } from "./gas";
 import { BroadcastMode } from "./lcdapi";
 import { Msg, MsgSend } from "./msgs";
 import { StdFee, StdTx } from "./types";
 import { OfflineSigner } from "./wallet";
 
 /**
- * Those fees are used by the higher level methods of SigningCosmosClient
+ * These fees are used by the higher level methods of SigningCosmosClient
  */
-export interface FeeTable {
+export interface CosmosFeeTable extends FeeTable {
   readonly send: StdFee;
 }
 
-const defaultFees: FeeTable = {
-  send: {
-    amount: coins(2000, "ucosm"),
-    gas: "80000", // 80k
-  },
-};
+const defaultGasPrice = GasPrice.fromString("0.025ucosm");
+const defaultGasLimits: GasLimits<CosmosFeeTable> = { send: 80000 };
+
+/** Use for testing only */
+export interface PrivateSigningCosmosClient {
+  readonly fees: CosmosFeeTable;
+}
 
 export class SigningCosmosClient extends CosmosClient {
   public readonly senderAddress: string;
 
   private readonly signer: OfflineSigner;
-  private readonly fees: FeeTable;
+  private readonly fees: CosmosFeeTable;
 
   /**
    * Creates a new client with signing capability to interact with a Cosmos SDK blockchain. This is the bigger brother of CosmosClient.
@@ -36,22 +38,23 @@ export class SigningCosmosClient extends CosmosClient {
    * @param apiUrl The URL of a Cosmos SDK light client daemon API (sometimes called REST server or REST API)
    * @param senderAddress The address that will sign and send transactions using this instance
    * @param signer An implementation of OfflineSigner which can provide signatures for transactions, potentially requiring user input.
-   * @param customFees The fees that are paid for transactions
+   * @param gasPrice The price paid per unit of gas
+   * @param gasLimits Custom overrides for gas limits related to specific transaction types
    * @param broadcastMode Defines at which point of the transaction processing the broadcastTx method returns
    */
   public constructor(
     apiUrl: string,
     senderAddress: string,
     signer: OfflineSigner,
-    customFees?: Partial<FeeTable>,
+    gasPrice: GasPrice = defaultGasPrice,
+    gasLimits: Partial<GasLimits<CosmosFeeTable>> = {},
     broadcastMode = BroadcastMode.Block,
   ) {
     super(apiUrl, broadcastMode);
     this.anyValidAddress = senderAddress;
-
     this.senderAddress = senderAddress;
     this.signer = signer;
-    this.fees = { ...defaultFees, ...(customFees || {}) };
+    this.fees = buildFeeTable<CosmosFeeTable>(gasPrice, defaultGasLimits, gasLimits);
   }
 
   public async getSequence(address?: string): Promise<GetSequenceResult> {